requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.1.1
annotated-types==0.5.0
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
cachetools==5.3.1
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
ctransformers==0.2.27
cycler==0.12.1
dataclasses-json==0.5.14
deprecation==2.1.0
entrypoints==0.4
faiss-cpu==1.7.4
Faker==22.0.0
favicon==0.7.0
filelock==3.12.4
fonttools==4.47.0
frozenlist==1.4.0
fsspec==2023.9.1
gitdb==4.0.10
GitPython==3.1.36
greenlet==2.0.2
h11==0.14.0
htbuilder==0.6.2
huggingface-hub==0.17.2
hydralit-components==1.0.10
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
langchain==0.0.295
langsmith==0.0.38
lxml==5.0.0
Markdown==3.5.1
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.2
mdurl==0.1.2
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.6
numpy==1.26.0
packaging==23.1
pandas==2.1.0
Pillow==9.5.0
protobuf==4.24.3
py-cpuinfo==9.0.0
pyarrow==13.0.0
pydantic==2.3.0
pydantic_core==2.6.3
pydeck==0.8.0
Pygments==2.16.1
pymdown-extensions==10.7
Pympler==1.0.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rich==13.5.3
rpds-py==0.10.3
safetensors==0.3.3
scikit-learn==1.3.0
scipy==1.11.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.1
soupsieve==2.5
SQLAlchemy==2.0.21
st-annotated-text==4.0.1
st-pages==0.4.5
streamlit==1.26.0
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.0
streamlit-embedcode==0.1.2
streamlit-extras==0.3.6
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.2
streamlit-modal==0.1.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
transformers==4.33.2
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.5
uvicorn==0.25.0
validators==0.22.0
watchdog==3.0.0
wrapt==1.15.0
yarl==1.9.2
zipp==3.17.0
openai==1.9.0
pdfminer.six==20231228
``` 
